+++ src/ifcblenderexport/blenderbim/bim/__init__.py
@@ -16,6 +16,7 @@ if bpy is not None:
         "cobie": None,
         "context": None,
         "covetool": None,
+        "csv": None,
         "debug": None,
         "geometry": None,
         "model": None,
@@ -140,11 +141,6 @@ if bpy is not None:
         operator.RemovePropertyTemplate,
         operator.AddSectionPlane,
         operator.RemoveSectionPlane,
-        operator.AddCsvAttribute,
-        operator.RemoveCsvAttribute,
-        operator.ExportIfcCsv,
-        operator.ImportIfcCsv,
-        operator.EyedropIfcCsv,
         operator.ReloadIfcFile,
         operator.AddIfcFile,
         operator.RemoveIfcFile,
@@ -255,7 +251,6 @@ if bpy is not None:
         ui.BIM_PT_document_information,
         ui.BIM_PT_constraints,
         ui.BIM_PT_search,
-        ui.BIM_PT_ifccsv,
         ui.BIM_PT_ifcclash,
         ui.BIM_PT_qa,
         ui.BIM_PT_library,
